+++ coresimd/src/x86/i586/avx2.rs
@@ -2887,15 +2887,6 @@ pub unsafe fn _mm256_extract_epi32(a: __m256i, imm8: i32) -> i32 {
     simd_extract(a.as_i32x8(), imm8)
 }
 
-/// Extract a 64-bit integer from `a`, selected with `imm8`.
-#[inline(always)]
-#[target_feature(enable = "avx2")]
-// This intrinsic has no corresponding instruction.
-pub unsafe fn _mm256_extract_epi64(a: __m256i, imm8: i32) -> i64 {
-    let imm8 = (imm8 & 3) as u32;
-    simd_extract(a.as_i64x4(), imm8)
-}
-
 /// Returns the first element of the input vector of [4 x double].
 #[inline(always)]
 #[target_feature(enable = "avx2")]
@@ -5244,13 +5235,6 @@ mod tests {
         let r2 = _mm256_extract_epi32(a, 11);
         assert_eq!(r1, -1);
         assert_eq!(r2, 3);
-    }
-
-    #[simd_test = "avx2"]
-    unsafe fn test_mm256_extract_epi64() {
-        let a = _mm256_setr_epi64x(0, 1, 2, 3);
-        let r = _mm256_extract_epi64(a, 3);
-        assert_eq!(r, 3);
     }
 
     #[simd_test = "avx2"]
